scan misc
