+++ android/src/main/res/values/strings.xml
@@ -113,4 +113,8 @@
 	<string name="update_metadata_keyword_metadata">Metadata</string>
 	<string name="scan_progress_bar_title">Multipart data</string>
 	<string name="scan_progress_bar_progress">%1$d / %2$d complete</string>
+    <string name="generic_ok">OK</string>
+	<string name="camera_jastification_title">Camera required!</string>
+	<string name="camera_jastification_message">To work with QR code we need camera permission, this is main functionality of this app!</string>
+	<string name="camera_no_permission_text">Camera permission not granted</string>
 </resources>
